src/requirements-dev.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
attrs==24.2.0; python_version >= '3.7'
black==24.4.2; python_version >= '3.8'
certifi==2024.7.4; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
coverage[toml]==7.5.4; python_version >= '3.8'
django==5.0.8; python_version >= '3.10'
django-stubs==5.0.2; python_version >= '3.8'
django-stubs-ext==5.0.4; python_version >= '3.8'
execnet==2.1.1; python_version >= '3.8'
flake8==7.1.0; python_full_version >= '3.8.1'
flake8-docstrings==1.7.0; python_version >= '3.7'
ghp-import==2.1.0
h11==0.14.0; python_version >= '3.7'
idna==3.7; python_version >= '3.5'
iniconfig==2.0.0; python_version >= '3.7'
isort==5.13.2; python_full_version >= '3.8.0'
jinja2==3.1.4; python_version >= '3.7'
markdown==3.6; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
mccabe==0.7.0; python_version >= '3.6'
mergedeep==1.3.4; python_version >= '3.6'
mkdocs==1.6.0; python_version >= '3.8'
mkdocs-get-deps==0.2.0; python_version >= '3.8'
mypy==1.10.1; python_version >= '3.8'
mypy-extensions==1.0.0; python_version >= '3.5'
outcome==1.3.0.post0; python_version >= '3.7'
packaging==24.1; python_version >= '3.8'
pathspec==0.12.1; python_version >= '3.8'
platformdirs==4.2.2; python_version >= '3.8'
pluggy==1.5.0; python_version >= '3.8'
pycodestyle==2.12.1; python_version >= '3.8'
pydocstyle==6.3.0; python_version >= '3.6'
pyflakes==3.2.0; python_version >= '3.8'
pysocks==1.7.1
pytest==8.2.2; python_version >= '3.8'
pytest-cov==5.0.0; python_version >= '3.8'
pytest-django==4.8.0; python_version >= '3.8'
pytest-sugar==1.0.0
pytest-xdist==3.6.1; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
pyyaml==6.0.2; python_version >= '3.8'
pyyaml-env-tag==0.1; python_version >= '3.6'
requests==2.32.3; python_version >= '3.8'
selenium==4.22.0; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
snowballstemmer==2.2.0
sortedcontainers==2.4.0
sqlparse==0.5.1; python_version >= '3.8'
termcolor==2.4.0; python_version >= '3.8'
trio==0.26.2; python_version >= '3.8'
trio-websocket==0.11.1; python_version >= '3.7'
types-pyyaml==6.0.12.20240808; python_version >= '3.8'
types-requests==2.32.0.20240622; python_version >= '3.8'
typing-extensions==4.12.2; python_version >= '3.8'
urllib3[socks]==2.2.2; python_version >= '3.8'
watchdog==4.0.1; python_version >= '3.8'
webdriver-manager==4.0.2; python_version >= '3.7'
websocket-client==1.8.0; python_version >= '3.8'
wsproto==1.2.0; python_full_version >= '3.7.0'
annotated-types==0.7.0; python_version >= '3.8'
django-ninja==1.2.1; python_version >= '3.7'
gunicorn==22.0.0; python_version >= '3.7'
markdown-it-py==3.0.0; python_version >= '3.8'
mdurl==0.1.2; python_version >= '3.7'
psycopg[binary]==3.2.1; python_version >= '3.8'
psycopg-binary==3.2.1
pydantic==2.8.2; python_version >= '3.8'
pydantic-core==2.20.1; python_version >= '3.8'
pygments==2.18.0; python_version >= '3.8'
rich==13.7.1; python_full_version >= '3.7.0'
uvicorn==0.30.1; python_version >= '3.8'
uvicorn-worker==0.2.0; python_version >= '3.8'
